compiler/src/main/java/yirgacheffe/compiler/statement/FunctionCall.java
package yirgacheffe.compiler.statement;

import org.objectweb.asm.MethodVisitor;
import org.objectweb.asm.Opcodes;
import yirgacheffe.compiler.expression.Expression;

public class FunctionCall implements Statement
{
	private Expression expression;

	public FunctionCall(Expression expression)
	{
		this.expression = expression;
	}

	@Override
	public void compile(MethodVisitor methodVisitor)
	{
		this.expression.compile(methodVisitor);

		int width = this.expression.getType().width();

		if (width == 1)
		{
			methodVisitor.visitInsn(Opcodes.POP);
		}
		else if (width == 2)
		{
			methodVisitor.visitInsn(Opcodes.POP2);
		}
	}
}
